Initial DCell::Node FSM
Finite state machines provide a great abstraction for building
asynchronous systems. This change begins to use Celluloid::FSM to model
the connection state of various nodes.

Knowing the connection state of nodes is essential to building robust
systems with DCell, because DCell provides an RPC mechanism on top of
what is assumed to be an unreliable and hostile network. With knowledge
of the underlying network state DCell can cancel RPCs made to downed
nodes so they don't hang forever.