Revert 9dc9245: Gossip protocol adapter
The gossip protocol was a neat feature! Unfortunately it never fully
worked right, and the abstractions DCell had in place weren't sufficient
to accomodate it properly.

This reverts the adapter and tries to put the codebase back on what was
working, stable code. Any attempts to add gossip protocols or other
built-in abstractions shouldn't reach out into the code so much and
should conform to a prespecified adapter API.

Lesson learned! Don't hit the merge button so easily